src/inference/pyproject.toml
[project]
dependencies = [
  "fastapi[standard]>=0.111.1",
  "uvicorn[standard]>=0.30.3",
  "Pillow==10.4.0",
  "opencv-python==4.10.0.84",
  "requests==2.32.3",
  "qoa4ml",
  "opentelemetry-distro>=0.48b0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0",
]

optional-dependencies.cpu = ["onnxruntime==1.18.1"]

name = "inference"
version = "0.1.0"
requires-python = "== 3.9.19"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["*.py"]
